Switch to a YAML-based configuration file format (#64)
This PR contains a backward-incompatible change changing the configuration file format from Python syntax to YAML syntax.
The new format tries to address some shortcomings of previous format: 
* it clearly differentiates between the task name and its parameters using different YAML fields (instead of having one-liner trying to encode all these things using forward slashes as separators);
* it removes the need of escaping (e.g. Linux paths containing forward slashes do not need to use pipe characters).
* it allows to use standard tools to validate the configuration file syntax against a schema (in this case the "yamale" library has been used which has been validated to work both on x86-64 and ARM platforms)

This PR is also:
* changing all handlers to use a new signature for their "handle()" function taking a list[str] for parameters instead of a single string potentially encoding a variable number of parameters;
* adding a counter "num_errors" of all exceptions captured during "handle()" function execution; this number of errors is reported over a dedicated MQTT topic as well and periodically on the log output;
* enhancing integration tests to check for the new "num_errors" parameters; the test is declared as successful only if ZERO errors are detected;

This change might be introducing regressions in some corner case due to the extensive code change.
Further improvements and further validation will follow in other PRs.

+++ integration_tests/psmqtt_container.py
@@ -5,31 +5,40 @@ import os
 import shutil
 import tempfile
 from testcontainers.core.container import DockerContainer
-from testcontainers.mqtt import MosquittoContainer
+from integration_tests.mosquitto_container import MosquittoContainerEnhanced
 
 class PSMQTTContainer(DockerContainer):
     """
     Specialization of DockerContainer to test PSMQTT docker image
     """
 
-    CONFIG_FILE = "integration-tests-psmqtt.conf"
+    CONFIG_FILE = "integration-tests-psmqtt.yaml"
 
-    def __init__(self, broker: MosquittoContainer) -> None:
-        super().__init__(image="psmqtt")
+    def __init__(self, broker: MosquittoContainerEnhanced) -> None:
+        super().__init__(image="psmqtt:latest")
 
         # IMPORTANT: to link with the MQTT broker we want to use the IP address internal to the docker network,
         #            and the standard MQTT port. The localhost:exposed_port address is not reachable from a
         #            docker container that has been started inside a docker network!
         broker_container = broker.get_wrapped_container()
         broker_ip = broker.get_docker_client().bridge_ip(broker_container.id)
-        broker_port = MosquittoContainer.MQTT_PORT
+        broker_port = MosquittoContainerEnhanced.MQTT_PORT
         print(f"Linking the {self.image} container with the MQTT broker at host:ip {broker_ip}:{broker_port}")
         config_file = self._prepare_config_file(broker_ip, broker_port)
 
         # bind-mount the generated, transient, config file to the standard location of config file within PSMQTT container
-        self.with_volume_mapping(config_file, "/opt/psmqtt/conf/psmqtt.conf", mode="ro")
+        self.with_volume_mapping(config_file, "/opt/psmqtt/conf/psmqtt.yaml", mode="ro")
 
-    def _prepare_config_file(self, broker_ip: str, broker_port: int):
+    def watch_for_internal_errors(self, broker: MosquittoContainerEnhanced) -> None:
+        # watch the psmqtt internal status topic
+        self.internal_mqtt_topics = {
+            "num_errors": self.get_mqtt_topic_prefix() + "/psmqtt_status/num_errors"
+            # add more in future?
+        }
+        print(f"Watching for internal errors in the PSMQTT container on topic: {self.internal_mqtt_topics['num_errors']}")
+        broker.watch_topics(list(self.internal_mqtt_topics.values()))
+
+    def _prepare_config_file(self, broker_ip: str, broker_port: int) -> str:
         TEST_DIR = os.path.dirname(os.path.abspath(__file__))
         original_cfgfile = os.path.join(TEST_DIR, self.CONFIG_FILE)
 
@@ -38,24 +47,39 @@ class PSMQTTContainer(DockerContainer):
 
         with open(temp_cfgfile, 'r+') as f:  # Open for reading and writing
             content = f.read()
+            original = content
             content = content.replace("__MQTT_BROKER_IP_PLACEHOLDER__", broker_ip)
             content = content.replace("__MQTT_BROKER_PORT_PLACEHOLDER__", f"{broker_port}")
-            f.seek(0)  # Rewind to the beginning
-            f.write(content)
-            f.truncate()  # Remove any remaining old content
+
+            if original != content:
+                f.seek(0)  # Rewind to the beginning
+                f.write(content)
+                f.truncate()  # Remove any remaining old content
+            else:
+                raise ValueError("Failed to replace placeholders in the configuration file")
 
         print(f"Prepared configuration file '{temp_cfgfile}' for use in integration tests")
         return temp_cfgfile
 
-    def is_running(self):
+    def is_running(self) -> bool:
         self.get_wrapped_container().reload()  # force refresh of container status
         # status = self.get_wrapped_container().attrs["State"]['Status']
         status = self.get_wrapped_container().status  # same as above
         return status == "running"
 
-    def get_short_id(self):
+    def get_short_id(self) -> str:
         # the docker container ID is a long string, the "short ID" is just the first 12 characters
         return self.get_wrapped_container().id[:12]
+
+    def get_mqtt_topic_prefix(self) -> str:
+        return f"psmqtt/{self.get_short_id()}"
+
+    def get_num_internal_errors(self, broker: MosquittoContainerEnhanced) -> int:
+        last_value = broker.get_last_payload_received_in_watched_topic(self.internal_mqtt_topics["num_errors"])
+        try:
+            return int(last_value)
+        except ValueError:
+            return -1
 
     def print_logs(self) -> str:
         print("** psmqtt LOGS [STDOUT]:")
